refactor: move format_verifier to utils
