+++ blarify/utils/format_verifier.py
@@ -1,4 +1,4 @@
 class FormatVerifier:
     @staticmethod
-    def is_path_uri(path) -> bool:
+    def is_path_uri(path: str) -> bool:
         return path.startswith("file://")
